Rename ability module to authz
