+++ examples/bri-3/src/bri/identity/bpiSubjects/api/subjects.controller.ts
@@ -9,7 +9,7 @@ import {
   Req,
 } from '@nestjs/common';
 import { CommandBus, QueryBus } from '@nestjs/cqrs';
-import { AbilityFactory } from '../../../ability/ability.factory';
+import { AbilityFactory } from '../../../authz/ability.factory';
 import { CreateBpiSubjectCommand } from '../capabilities/createBpiSubject/createBpiSubject.command';
 import { DeleteBpiSubjectCommand } from '../capabilities/deleteBpiSubject/deleteBpiSubject.command';
 import { GetAllBpiSubjectsQuery } from '../capabilities/getAllBpiSubjects/getAllBpiSubjects.query';
